+++ lib/src/saf/saf.dart
@@ -3,17 +3,19 @@ import 'dart:typed_data';
 
 import '../../saf.dart';
 import '../channels.dart';
+import '../common/functional_extender.dart';
 import 'common.dart';
 
 /// {@template sharedstorage.saf.openDocumentTree}
 /// Start Activity Action: Allow the user to pick a directory subtree.
+///
 /// When invoked, the system will display the various `DocumentsProvider`
 /// instances installed on the device, letting the user navigate through them.
 /// Apps can fully manage documents within the returned directory.
 ///
-/// [Refer to details](https://developer.android.com/reference/android/content/Intent#ACTION_OPEN_DOCUMENT_TREE)
+/// [Refer to details](https://developer.android.com/reference/android/content/Intent#ACTION_OPEN_DOCUMENT_TREE).
 ///
-/// support the initial directory of the directory picker
+/// support the initial directory of the directory picker.
 /// {@endtemplate}
 Future<Uri?> openDocumentTree({
   bool grantWritePermission = true,
@@ -29,55 +31,48 @@ Future<Uri?> openDocumentTree({
   final selectedDirectoryUri =
       await kDocumentFileChannel.invokeMethod<String?>(kOpenDocumentTree, args);
 
-  if (selectedDirectoryUri == null) return null;
-
-  return Uri.parse(selectedDirectoryUri);
+  return selectedDirectoryUri?.apply((e) => Uri.parse(e));
 }
 
 /// {@template sharedstorage.saf.persistedUriPermissions}
 /// Returns an `List<Uri>` with all persisted [Uri]
 ///
-/// To persist an [Uri] call `openDocumentTree`
-/// and to remove an persisted [Uri] call `releasePersistableUriPermission`
+/// To persist an [Uri] call `openDocumentTree`.
+///
+/// To remove an persisted [Uri] call `releasePersistableUriPermission`.
 /// {@endtemplate}
 Future<List<UriPermission>?> persistedUriPermissions() async {
-  const kPersistedUriPermissions = 'persistedUriPermissions';
-
   final persistedUriPermissions =
-      await kDocumentFileChannel.invokeListMethod(kPersistedUriPermissions);
+      await kDocumentFileChannel.invokeListMethod('persistedUriPermissions');
 
-  if (persistedUriPermissions == null) return null;
-
-  return persistedUriPermissions
-      .map((e) => UriPermission.fromMap(Map.from(e as Map)))
-      .toList();
+  return persistedUriPermissions?.apply(
+    (p) => p.map((e) => UriPermission.fromMap(Map.from(e as Map))).toList(),
+  );
 }
 
 /// {@template sharedstorage.saf.releasePersistableUriPermission}
-/// Will revoke an persistable Uri
+/// Will revoke an persistable Uri.
 ///
 /// Call this when your App no longer wants the permission of an [Uri] returned
-/// by `openDocumentTree` method
+/// by `openDocumentTree` method.
 ///
-/// To get the current persisted [Uri]s call `persistedUriPermissions`
+/// To get the current persisted [Uri]s call `persistedUriPermissions`.
+///
+/// [Refer to details](https://developer.android.com/reference/android/content/ContentResolver#releasePersistableUriPermission(android.net.Uri,%20int)).
 /// {@endtemplate}
 Future<void> releasePersistableUriPermission(Uri directory) async {
-  const kReleasePersistableUriPermission = 'releasePersistableUriPermission';
-
-  final args = <String, String>{'uri': '$directory'};
-
   await kDocumentFileChannel.invokeMethod(
-    kReleasePersistableUriPermission,
-    args,
+    'releasePersistableUriPermission',
+    <String, String>{'uri': '$directory'},
   );
 }
 
 /// {@template sharedstorage.saf.isPersistedUri}
-/// Convenient method to verify if a given [uri]
-/// is allowed to be write or read from SAF API's
+/// Convenient method to verify if a given [uri].
+/// is allowed to be write or read from SAF API's.
 ///
 /// This uses the `releasePersistableUriPermission` method to get the List
-/// of allowed [Uri]s then will verify if the [uri] is included in
+/// of allowed [Uri]s then will verify if the [uri] is included in.
 /// {@endtemplate}
 Future<bool> isPersistedUri(Uri uri) async {
   final persistedUris = await persistedUriPermissions();
@@ -86,35 +81,25 @@ Future<bool> isPersistedUri(Uri uri) async {
 }
 
 /// {@template sharedstorage.saf.canRead}
-/// Equivalent to `DocumentFile.canRead`
+/// Equivalent to `DocumentFile.canRead`.
 ///
-/// [Refer to details](https://developer.android.com/reference/androidx/documentfile/provider/DocumentFile#canRead())
+/// [Refer to details](https://developer.android.com/reference/androidx/documentfile/provider/DocumentFile#canRead()).
 /// {@endtemplate}
-Future<bool?> canRead(Uri uri) async {
-  const kCanRead = 'canRead';
-
-  final args = {'uri': '$uri'};
-
-  return kDocumentFileChannel.invokeMethod<bool>(kCanRead, args);
-}
+Future<bool?> canRead(Uri uri) async => kDocumentFileChannel
+    .invokeMethod<bool>('canRead', <String, String>{'uri': '$uri'});
 
 /// {@template sharedstorage.saf.canWrite}
-/// Equivalent to `DocumentFile.canWrite`
+/// Equivalent to `DocumentFile.canWrite`.
 ///
-/// [Refer to details](https://developer.android.com/reference/androidx/documentfile/provider/DocumentFile#canWrite())
+/// [Refer to details](https://developer.android.com/reference/androidx/documentfile/provider/DocumentFile#canWrite()).
 /// {@endtemplate}
-Future<bool?> canWrite(Uri uri) async {
-  const kCanWrite = 'canWrite';
-
-  final args = {'uri': '$uri'};
-
-  return kDocumentFileChannel.invokeMethod<bool>(kCanWrite, args);
-}
+Future<bool?> canWrite(Uri uri) async => kDocumentFileChannel
+    .invokeMethod<bool>('canWrite', <String, String>{'uri': '$uri'});
 
 /// {@template sharedstorage.saf.getDocumentThumbnail}
-/// Equivalent to `DocumentsContract.getDocumentThumbnail`
+/// Equivalent to `DocumentsContract.getDocumentThumbnail`.
 ///
-/// [Refer to details](https://developer.android.com/reference/android/provider/DocumentsContract#getDocumentThumbnail(android.content.ContentResolver,%20android.net.Uri,%20android.graphics.Point,%20android.os.CancellationSignal))
+/// [Refer to details](https://developer.android.com/reference/android/provider/DocumentsContract#getDocumentThumbnail(android.content.ContentResolver,%20android.net.Uri,%20android.graphics.Point,%20android.os.CancellationSignal)).
 /// {@endtemplate}
 Future<DocumentBitmap?> getDocumentThumbnail({
   required Uri rootUri,
@@ -122,8 +107,6 @@ Future<DocumentBitmap?> getDocumentThumbnail({
   required double width,
   required double height,
 }) async {
-  const kGetDocumentThumbnail = 'getDocumentThumbnail';
-
   final args = <String, dynamic>{
     'rootUri': '$rootUri',
     'documentId': documentId,
@@ -132,19 +115,17 @@ Future<DocumentBitmap?> getDocumentThumbnail({
   };
 
   final bitmap = await kDocumentsContractChannel
-      .invokeMapMethod<String, dynamic>(kGetDocumentThumbnail, args);
+      .invokeMapMethod<String, dynamic>('getDocumentThumbnail', args);
 
-  if (bitmap == null) return null;
-
-  return DocumentBitmap.fromMap(bitmap);
+  return bitmap?.apply((b) => DocumentBitmap.fromMap(b));
 }
 
 /// {@template sharedstorage.saf.listFiles}
-/// **Important**: Ensure you have read permission by calling `canRead` before calling `listFiles`
+/// **Important**: Ensure you have read permission by calling `canRead` before calling `listFiles`.
 ///
-/// Emits a new event for each child document file
+/// Emits a new event for each child document file.
 ///
-/// Works with small and large data file sets
+/// Works with small and large data file sets.
 ///
 /// ```dart
 /// /// Usage:
@@ -162,17 +143,15 @@ Future<DocumentBitmap?> getDocumentThumbnail({
 /// });
 /// ```
 ///
-/// [Refer to details](https://stackoverflow.com/questions/41096332/issues-traversing-through-directory-hierarchy-with-android-storage-access-framew)
+/// [Refer to details](https://stackoverflow.com/questions/41096332/issues-traversing-through-directory-hierarchy-with-android-storage-access-framew).
 /// {@endtemplate}
 Stream<PartialDocumentFile> listFiles(
   Uri uri, {
   required List<DocumentFileColumn> columns,
 }) {
-  const kListFiles = 'listFiles';
-
   final args = <String, dynamic>{
     'uri': '$uri',
-    'event': kListFiles,
+    'event': 'listFiles',
     'columns': columns.map((e) => '$e').toList(),
   };
 
@@ -185,130 +164,108 @@ Stream<PartialDocumentFile> listFiles(
 }
 
 /// {@template sharedstorage.saf.exists}
-/// Verify if a given [uri] exists
+///  Equivalent to `DocumentFile.exists`.
+///
+/// Verify wheter or not a given [uri] exists.
+///
+/// [Refer to details](https://developer.android.com/reference/androidx/documentfile/provider/DocumentFile#exists()).
 /// {@endtemplate}
-Future<bool?> exists(Uri uri) async {
-  const kExists = 'exists';
-
-  final args = <String, String>{'uri': '$uri'};
-
-  return kDocumentFileChannel.invokeMethod<bool>(kExists, args);
-}
+Future<bool?> exists(Uri uri) async => kDocumentFileChannel
+    .invokeMethod<bool>('exists', <String, String>{'uri': '$uri'});
 
 /// {@template sharedstorage.saf.buildDocumentUriUsingTree}
-/// Equivalent to `DocumentsContract.buildDocumentUriUsingTree`
+/// Equivalent to `DocumentsContract.buildDocumentUriUsingTree`.
 ///
-/// [Refer to details](https://developer.android.com/reference/android/provider/DocumentsContract#buildDocumentUriUsingTree%28android.net.Uri,%20java.lang.String%29)
+/// [Refer to details](https://developer.android.com/reference/android/provider/DocumentsContract#buildDocumentUriUsingTree%28android.net.Uri,%20java.lang.String%29).
 /// {@endtemplate}
 Future<Uri?> buildDocumentUriUsingTree(Uri treeUri, String documentId) async {
-  const kBuildDocumentUriUsingTree = 'buildDocumentUriUsingTree';
-
   final args = <String, String>{
     'treeUri': '$treeUri',
     'documentId': documentId,
   };
 
   final uri = await kDocumentsContractChannel.invokeMethod<String>(
-    kBuildDocumentUriUsingTree,
+    'buildDocumentUriUsingTree',
     args,
   );
 
-  if (uri == null) return null;
-
-  return Uri.parse(uri);
+  return uri?.apply((u) => Uri.parse(u));
 }
 
 /// {@template sharedstorage.saf.buildDocumentUri}
-/// Equivalent to `DocumentsContract.buildDocumentUri`
+/// Equivalent to `DocumentsContract.buildDocumentUri`.
 ///
-/// [Refer to details](https://developer.android.com/reference/android/provider/DocumentsContract#buildDocumentUri%28java.lang.String,%20java.lang.String%29)
+/// [Refer to details](https://developer.android.com/reference/android/provider/DocumentsContract#buildDocumentUri%28java.lang.String,%20java.lang.String%29).
 /// {@endtemplate}
 Future<Uri?> buildDocumentUri(String authority, String documentId) async {
-  const kBuildDocumentUri = 'buildDocumentUri';
-
   final args = <String, String>{
     'authority': authority,
     'documentId': documentId,
   };
 
   final uri = await kDocumentsContractChannel.invokeMethod<String>(
-    kBuildDocumentUri,
+    'buildDocumentUri',
     args,
   );
 
-  if (uri == null) return null;
-
-  return Uri.parse(uri);
+  return uri?.apply((u) => Uri.parse(u));
 }
 
 /// {@template sharedstorage.saf.buildDocumentUri}
-/// Equivalent to `DocumentsContract.buildDocumentUri`
+/// Equivalent to `DocumentsContract.buildDocumentUri`.
 ///
-/// [Refer to details](https://developer.android.com/reference/android/provider/DocumentsContract#buildDocumentUri%28java.lang.String,%20java.lang.String%29)
+/// [Refer to details](https://developer.android.com/reference/android/provider/DocumentsContract#buildDocumentUri%28java.lang.String,%20java.lang.String%29).
 /// {@endtemplate}
 Future<Uri?> buildTreeDocumentUri(String authority, String documentId) async {
-  const kBuildTreeDocumentUri = 'buildTreeDocumentUri';
-
   final args = <String, String>{
     'authority': authority,
     'documentId': documentId,
   };
 
   final uri = await kDocumentsContractChannel.invokeMethod<String>(
-    kBuildTreeDocumentUri,
+    'buildTreeDocumentUri',
     args,
   );
 
-  if (uri == null) return null;
-
-  return Uri.parse(uri);
+  return uri?.apply((u) => Uri.parse(u));
 }
 
 /// {@template sharedstorage.saf.delete}
-/// Equivalent to `DocumentFile.delete`
+/// Equivalent to `DocumentFile.delete`.
 ///
-/// Returns `true` if deleted successfully
+/// Returns `true` if deleted successfully.
 ///
-/// [Refer to details](https://developer.android.com/reference/androidx/documentfile/provider/DocumentFile#delete%28%29)
+/// [Refer to details](https://developer.android.com/reference/androidx/documentfile/provider/DocumentFile#delete%28%29).
 /// {@endtemplate}
-Future<bool?> delete(Uri uri) async {
-  const kDelete = 'delete';
-
-  final args = <String, String>{'uri': '$uri'};
-
-  return kDocumentFileChannel.invokeMethod<bool>(kDelete, args);
-}
+Future<bool?> delete(Uri uri) async => kDocumentFileChannel
+    .invokeMethod<bool>('delete', <String, String>{'uri': '$uri'});
 
 /// {@template sharedstorage.saf.createDirectory}
-/// Create a direct child document tree named `displayName` given a parent `parentUri`
+/// Create a direct child document tree named `displayName` given a parent `parentUri`.
 ///
-/// Equivalent to `DocumentFile.createDirectory`
+/// Equivalent to `DocumentFile.createDirectory`.
 ///
-/// [Refer to details](https://developer.android.com/reference/androidx/documentfile/provider/DocumentFile#createDirectory%28java.lang.String%29)
+/// [Refer to details](https://developer.android.com/reference/androidx/documentfile/provider/DocumentFile#createDirectory%28java.lang.String%29).
 /// {@endtemplate}
 Future<DocumentFile?> createDirectory(Uri parentUri, String displayName) async {
-  const kCreateDirectory = 'createDirectory';
-
   final args = <String, String>{
     'uri': '$parentUri',
     'displayName': displayName,
   };
 
   final createdDocumentFile = await kDocumentFileChannel
-      .invokeMapMethod<String, dynamic>(kCreateDirectory, args);
+      .invokeMapMethod<String, dynamic>('createDirectory', args);
 
-  if (createdDocumentFile == null) return null;
-
-  return DocumentFile.fromMap(createdDocumentFile);
+  return createdDocumentFile?.apply((c) => DocumentFile.fromMap(c));
 }
 
 /// {@template sharedstorage.saf.createFile}
-/// Convenient method to create files using either String or raw bytes
+/// Convenient method to create files using either String or raw bytes.
 ///
 /// Under the hood this method calls `createFileAsString` or `createFileAsBytes`
-/// depending on which argument is passed
+/// depending on which argument is passed.
 ///
-/// If both (bytes and content) are passed, the bytes will be used and the content will be ignored
+/// If both (bytes and content) are passed, the bytes will be used and the content will be ignored.
 /// {@endtemplate}
 Future<DocumentFile?> createFile(
   Uri parentUri, {
@@ -338,12 +295,14 @@ Future<DocumentFile?> createFile(
 }
 
 /// {@template sharedstorage.saf.createFileAsBytes}
-/// Create a direct child document of `parentUri`
-/// - `mimeType` is the type of document following [this specs](https://www.iana.org/assignments/media-types/media-types.xhtml)
-/// - `displayName` is the name of the document, must be a valid file name
-/// - `bytes` is the content of the document as a list of bytes `Uint8List`
+/// Create a direct child document of `parentUri`.
+/// - `mimeType` is the type of document following [this specs](https://www.iana.org/assignments/media-types/media-types.xhtml).
+/// - `displayName` is the name of the document, must be a valid file name.
+/// - `bytes` is the content of the document as a list of bytes `Uint8List`.
 ///
-/// Returns the created file as a `DocumentFile`
+/// Returns the created file as a `DocumentFile`.
+///
+/// Mirror of [`DocumentFile.createFile`](https://developer.android.com/reference/androidx/documentfile/provider/DocumentFile#createFile(java.lang.String,%20java.lang.String))
 /// {@endtemplate}
 Future<DocumentFile?> createFileAsBytes(
   Uri parentUri, {
@@ -351,8 +310,6 @@ Future<DocumentFile?> createFileAsBytes(
   required String displayName,
   required Uint8List bytes,
 }) async {
-  const kCreateFile = 'createFile';
-
   final directoryUri = '$parentUri';
 
   final args = <String, dynamic>{
@@ -362,12 +319,12 @@ Future<DocumentFile?> createFileAsBytes(
     'directoryUri': directoryUri,
   };
 
-  return invokeMapMethod(kCreateFile, args);
+  return invokeMapMethod('createFile', args);
 }
 
 /// {@template sharedstorage.saf.createFileAsString}
-/// Convenient method to create a file
-/// using `content` as String instead Uint8List
+/// Convenient method to create a file.
+/// using `content` as String instead Uint8List.
 /// {@endtemplate}
 Future<DocumentFile?> createFileAsString(
   Uri parentUri, {
@@ -384,190 +341,158 @@ Future<DocumentFile?> createFileAsString(
 }
 
 /// {@template sharedstorage.saf.length}
-/// Equivalent to `DocumentFile.length`
+/// Equivalent to `DocumentFile.length`.
 ///
-/// Returns the size of a given document `uri` in bytes
+/// Returns the size of a given document `uri` in bytes.
 ///
-/// [Refer to details](https://developer.android.com/reference/androidx/documentfile/provider/DocumentFile#length%28%29)
+/// [Refer to details](https://developer.android.com/reference/androidx/documentfile/provider/DocumentFile#length%28%29).
 /// {@endtemplate}
-Future<int?> documentLength(Uri uri) async {
-  const kLength = 'length';
-
-  final args = <String, String>{'uri': '$uri'};
-
-  return kDocumentFileChannel.invokeMethod<int>(kLength, args);
-}
+Future<int?> documentLength(Uri uri) async => kDocumentFileChannel
+    .invokeMethod<int>('length', <String, String>{'uri': '$uri'});
 
 /// {@template sharedstorage.saf.lastModified}
-/// Equivalent to `DocumentFile.lastModified`
+/// Equivalent to `DocumentFile.lastModified`.
 ///
-/// [Refer to details](https://developer.android.com/reference/androidx/documentfile/provider/DocumentFile#lastModified%28%29)
+/// [Refer to details](https://developer.android.com/reference/androidx/documentfile/provider/DocumentFile#lastModified%28%29).
 /// {@endtemplate}
 Future<DateTime?> lastModified(Uri uri) async {
   const kLastModified = 'lastModified';
 
-  final args = <String, String>{'uri': '$uri'};
+  final inMillisecondsSinceEpoch = await kDocumentFileChannel
+      .invokeMethod<int>(kLastModified, <String, String>{'uri': '$uri'});
 
-  final inMillisecondsSinceEpoch =
-      await kDocumentFileChannel.invokeMethod<int>(kLastModified, args);
-
-  if (inMillisecondsSinceEpoch == null) return null;
-
-  return DateTime.fromMillisecondsSinceEpoch(inMillisecondsSinceEpoch);
+  return inMillisecondsSinceEpoch
+      ?.takeIf((i) => i > 0)
+      ?.apply((i) => DateTime.fromMillisecondsSinceEpoch(i));
 }
 
 /// {@template sharedstorage.saf.findFile}
-/// Equivalent to `DocumentFile.findFile`
+/// Equivalent to `DocumentFile.findFile`.
 ///
-/// If you want to check if a given document file exists by their [displayName] prefer using `child` instead
+/// If you want to check if a given document file exists by their [displayName] prefer using `child` instead.
 ///
-/// [Refer to details](https://developer.android.com/reference/androidx/documentfile/provider/DocumentFile#findFile%28java.lang.String%29)
+/// [Refer to details](https://developer.android.com/reference/androidx/documentfile/provider/DocumentFile#findFile%28java.lang.String%29).
 /// {@endtemplate}
 Future<DocumentFile?> findFile(Uri directoryUri, String displayName) async {
-  const kFindFile = 'findFile';
-
   final args = <String, String>{
     'uri': '$directoryUri',
     'displayName': displayName,
   };
 
-  return invokeMapMethod(kFindFile, args);
+  return invokeMapMethod('findFile', args);
 }
 
 /// {@template sharedstorage.saf.renameTo}
-/// Rename the current document `uri` to a new `displayName`
+/// Rename the current document `uri` to a new `displayName`.
 ///
 /// **Note: after using this method `uri` is not longer valid,
-/// use the returned document instead**
+/// use the returned document instead**.
 ///
-/// Returns the updated document
+/// Returns the updated document.
 ///
-/// Equivalent to `DocumentFile.renameTo`
+/// Equivalent to `DocumentFile.renameTo`.
 ///
-/// [Refer to details](https://developer.android.com/reference/androidx/documentfile/provider/DocumentFile#renameTo%28java.lang.String%29)
+/// [Refer to details](https://developer.android.com/reference/androidx/documentfile/provider/DocumentFile#renameTo%28java.lang.String%29).
 /// {@endtemplate}
-Future<DocumentFile?> renameTo(Uri uri, String displayName) async {
-  const kRenameTo = 'renameTo';
-
-  final args = <String, String>{'uri': '$uri', 'displayName': displayName};
-
-  return invokeMapMethod(kRenameTo, args);
-}
+Future<DocumentFile?> renameTo(Uri uri, String displayName) async =>
+    invokeMapMethod(
+      'renameTo',
+      <String, String>{'uri': '$uri', 'displayName': displayName},
+    );
 
 /// {@template sharedstorage.saf.fromTreeUri}
-/// Create a new `DocumentFile` instance given `uri`
+/// Create a new `DocumentFile` instance given `uri`.
 ///
-/// Equivalent to `DocumentFile.fromTreeUri`
+/// Equivalent to `DocumentFile.fromTreeUri`.
 ///
-/// [Refer to details](https://developer.android.com/reference/androidx/documentfile/provider/DocumentFile#fromTreeUri%28android.content.Context,%20android.net.Uri%29)
+/// [Refer to details](https://developer.android.com/reference/androidx/documentfile/provider/DocumentFile#fromTreeUri%28android.content.Context,%20android.net.Uri%29).
 /// {@endtemplate}
-Future<DocumentFile?> fromTreeUri(Uri uri) async {
-  const kFromTreeUri = 'fromTreeUri';
-
-  final args = <String, String>{'uri': '$uri'};
-
-  return invokeMapMethod(kFromTreeUri, args);
-}
+Future<DocumentFile?> fromTreeUri(Uri uri) async =>
+    invokeMapMethod('fromTreeUri', <String, String>{'uri': '$uri'});
 
 /// {@template sharedstorage.saf.child}
-/// Return the `child` of the given `uri` if it exists otherwise `null`
+/// Return the `child` of the given `uri` if it exists otherwise `null`.
 ///
 /// It's faster than [DocumentFile.findFile]
 /// `path` is the single file name or file path. Empty string returns to itself.
 ///
-/// Equivalent to `DocumentFile.child` extension/overload
+/// Equivalent to `DocumentFile.child` extension/overload.
 ///
 /// [Refer to details](https://developer.android.com/reference/androidx/documentfile/provider/DocumentFile#fromTreeUri%28android.content.Context,%20android.net.Uri%29)
 /// {@endtemplate}
+@willbemovedsoon
 Future<DocumentFile?> child(
   Uri uri,
   String path, {
   bool requiresWriteAccess = false,
 }) async {
-  const kChild = 'child';
-
   final args = <String, dynamic>{
     'uri': '$uri',
     'path': path,
     'requiresWriteAccess': requiresWriteAccess,
   };
 
-  return invokeMapMethod(kChild, args);
+  return invokeMapMethod('child', args);
 }
 
 /// {@template sharedstorage.saf.openDocumentFile}
-/// It's a convenience method to launch the default application associated with the given MIME type
-/// and can't be considered an official SAF API.
+/// It's a convenience method to launch the default application associated
+/// with the given MIME type and can't be considered an official SAF API.
 ///
-/// Launch `ACTION_VIEW` intent to open the given document `uri`
+/// Launch `ACTION_VIEW` intent to open the given document `uri`.
 ///
 /// Throws an `PlatformException` with code `EXCEPTION_ACTIVITY_NOT_FOUND` if the activity is not found
 /// to the respective MIME type of the give Uri.
 ///
-/// Returns `true` if launched successfully otherwise `false`
+/// Returns `true` if launched successfully otherwise `false`.
 /// {@endtemplate}
 Future<bool?> openDocumentFile(Uri uri) async {
-  const kOpenDocumentFile = 'openDocumentFile';
-
-  final args = <String, String>{'uri': '$uri'};
-
-  final launched = await kDocumentFileHelperChannel.invokeMethod<bool>(
-    kOpenDocumentFile,
-    args,
+  final successfullyLaunched =
+      await kDocumentFileHelperChannel.invokeMethod<bool>(
+    'openDocumentFile',
+    <String, String>{'uri': '$uri'},
   );
 
-  return launched;
+  return successfullyLaunched;
 }
 
 /// {@template sharedstorage.saf.parentFile}
-/// Get the parent file of the given `uri`
+/// Get the parent file of the given `uri`.
 ///
-/// Equivalent to `DocumentFile.getParentFile`
+/// Equivalent to `DocumentFile.getParentFile`.
 ///
-/// [Refer to details](https://developer.android.com/reference/androidx/documentfile/provider/DocumentFile#getParentFile%28%29)
+/// [Refer to details](https://developer.android.com/reference/androidx/documentfile/provider/DocumentFile#getParentFile%28%29).
 /// {@endtemplate}
-Future<DocumentFile?> parentFile(Uri uri) async {
-  const kParentFile = 'parentFile';
-
-  final args = <String, String>{'uri': '$uri'};
-
-  return invokeMapMethod(kParentFile, args);
-}
+Future<DocumentFile?> parentFile(Uri uri) async =>
+    invokeMapMethod('parentFile', <String, String>{'uri': '$uri'});
 
 /// {@template sharedstorage.saf.copy}
-/// Copy a document `uri` to the `destination`
+/// Copy a document `uri` to the `destination`.
 ///
-/// This API uses the `createFile` and `getDocumentContent` API's behind the scenes
+/// This API uses the `createFile` and `getDocumentContent` API's behind the scenes.
 /// {@endtemplate}
 Future<DocumentFile?> copy(Uri uri, Uri destination) async {
-  const kCopy = 'copy';
-
   final args = <String, String>{'uri': '$uri', 'destination': '$destination'};
 
-  return invokeMapMethod(kCopy, args);
+  return invokeMapMethod('copy', args);
 }
 
 /// {@template sharedstorage.saf.getDocumentContent}
-/// Get content of a given document `uri`
+/// Get content of a given document `uri`.
 ///
-/// Equivalent to `contentDescriptor` usage
+/// Equivalent to `contentDescriptor` usage.
 ///
-/// [Refer to details](https://developer.android.com/training/data-storage/shared/documents-files#input_stream)
+/// [Refer to details](https://developer.android.com/training/data-storage/shared/documents-files#input_stream).
 /// {@endtemplate}
-Future<Uint8List?> getDocumentContent(Uri uri) async {
-  const kGetDocumentContent = 'getDocumentContent';
-
-  final args = <String, String>{'uri': '$uri'};
-
-  return kDocumentFileChannel.invokeMethod<Uint8List>(
-    kGetDocumentContent,
-    args,
-  );
-}
+Future<Uint8List?> getDocumentContent(Uri uri) async =>
+    kDocumentFileChannel.invokeMethod<Uint8List>(
+      'getDocumentContent',
+      <String, String>{'uri': '$uri'},
+    );
 
 /// {@template sharedstorage.saf.getDocumentContentAsString}
 /// Helper method to read document using
-/// `getDocumentContent` and get the content as String instead as `Uint8List`
+/// `getDocumentContent` and get the content as String instead as `Uint8List`.
 /// {@endtemplate}
 Future<String?> getDocumentContentAsString(
   Uri uri, {
@@ -575,21 +500,14 @@ Future<String?> getDocumentContentAsString(
 }) async {
   final bytes = await getDocumentContent(uri);
 
-  if (bytes == null) return null;
-
-  return String.fromCharCodes(bytes);
+  return bytes?.apply((a) => String.fromCharCodes(a));
 }
 
 /// {@template sharedstorage.saf.getDocumentContentAsString}
 /// Helper method to generate the file path of the given `uri`
 ///
 /// See [Get real path from URI, Android KitKat new storage access framework](https://stackoverflow.com/questions/20067508/get-real-path-from-uri-android-kitkat-new-storage-access-framework/20559175#20559175)
-/// for details
+/// for details.
 /// {@endtemplate}
-Future<String?> getRealPathFromUri(Uri uri) async {
-  const kGetRealPathFromUri = 'getRealPathFromUri';
-
-  final args = <String, String>{'uri': '$uri'};
-
-  return kDocumentFileHelperChannel.invokeMethod(kGetRealPathFromUri, args);
-}
+Future<String?> getRealPathFromUri(Uri uri) async => kDocumentFileHelperChannel
+    .invokeMethod('getRealPathFromUri', <String, String>{'uri': '$uri'});